+++ src/components/Hero/Hero.stories.tsx
@@ -9,6 +9,7 @@ import Heading from "../Heading/Heading";
 import Hero, { heroSecondaryTypes, heroTypesArray } from "./Hero";
 import SimpleGrid from "../Grid/SimpleGrid";
 import Text from "../Text/Text";
+import { getPlaceholderImage } from "../../utils/utils";
 
 const secondarySubHeaderText = (
   <>
@@ -55,7 +56,7 @@ const otherSubHeaderTextLong = (
 );
 const imageProps = {
   alt: "Image example",
-  src: "//placekitten.com/800/400",
+  src: getPlaceholderImage(),
 };
 
 const meta: Meta<typeof Hero> = {
@@ -108,7 +109,7 @@ export const WithControls: Story = {
     (args.heroType === "primary" && (
       <Hero
         {...args}
-        backgroundImageSrc="//placekitten.com/2400/800"
+        backgroundImageSrc={getPlaceholderImage()}
         heading={
           <Heading
             level="h1"
@@ -156,7 +157,7 @@ export const WithControls: Story = {
     (args.heroType === "campaign" && (
       <Hero
         {...args}
-        backgroundImageSrc="//placekitten.com/2400/800"
+        backgroundImageSrc={getPlaceholderImage()}
         heading={<Heading level="h1" id="1" text="Hero Campaign" />}
         heroType={args.heroType}
         imageProps={args.imageProps}
@@ -169,7 +170,7 @@ export const WithControls: Story = {
         heroType={args.heroType}
         imageProps={{
           ...args.imageProps,
-          src: "//placekitten.com/1200/400",
+          src: getPlaceholderImage(),
         }}
         subHeaderText={
           <>
@@ -191,7 +192,7 @@ export const WithControls: Story = {
 export const Primary: Story = {
   render: () => (
     <Hero
-      backgroundImageSrc="//placekitten.com/1600/800"
+      backgroundImageSrc={getPlaceholderImage()}
       heading={
         <Heading
           id="primary-hero"
@@ -262,7 +263,7 @@ export const Campaign: Story = {
           text="Campaign Hero at Default Height"
         />
         <Hero
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc={getPlaceholderImage()}
           heroType="campaign"
           heading={
             <Heading
@@ -282,7 +283,7 @@ export const Campaign: Story = {
           text="Campaign Hero with Long Text"
         />
         <Hero
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc={getPlaceholderImage()}
           heroType="campaign"
           heading={
             <Heading
@@ -305,7 +306,7 @@ export const CampaignDarkBackgroundImage: Story = {
     <Stack spacing="l">
       <div>
         <Hero
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc={getPlaceholderImage()}
           heroType="campaign"
           heading={
             <Heading