+++ src/components/Hero/Hero.test.tsx
@@ -4,6 +4,7 @@ import * as React from "react";
 import renderer from "react-test-renderer";
 import Heading from "../Heading/Heading";
 import Hero from "./Hero";
+import { getPlaceholderImage } from "../../utils/utils";
 
 export const subHeaderText = (
   <>
@@ -19,7 +20,7 @@ export const otherSubHeaderText =
   "Visit us today.";
 const imageProps = {
   alt: "Image example",
-  src: "//placekitten.com/800/400",
+  src: getPlaceholderImage("smaller", 0),
 };
 
 describe("Hero accessbility tests", () => {
@@ -29,7 +30,7 @@ describe("Hero accessbility tests", () => {
         heroType="primary"
         heading={<Heading level="h1" id="a11y-hero" text="Hero Primary" />}
         subHeaderText="Example Subtitle"
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
       />
     );
     expect(await axe(container)).toHaveNoViolations();
@@ -61,7 +62,7 @@ describe("Hero accessbility tests", () => {
   it("passes for type Campaign", async () => {
     const { container } = render(
       <Hero
-        backgroundImageSrc="//placekitten.com/g/2400/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="campaign"
         heading={<Heading level="h1" id="a11y-hero" text="Hero Campaign" />}
         imageProps={imageProps}
@@ -90,7 +91,7 @@ describe("Hero", () => {
         heroType="primary"
         heading={<Heading level="h1" id="primary-hero" text="Hero Primary" />}
         subHeaderText="Example Subtitle"
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
       />
     );
 
@@ -98,7 +99,7 @@ describe("Hero", () => {
     expect(screen.getByText("Example Subtitle")).toBeInTheDocument();
     expect(screen.getByTestId("hero")).toHaveAttribute(
       "style",
-      "background-image: url(//placekitten.com/1600/800);"
+      `background-image: url(${getPlaceholderImage("smaller", 0)});`
     );
   });
 
@@ -119,7 +120,7 @@ describe("Hero", () => {
     expect(screen.getByRole("img")).toBeInTheDocument();
     expect(screen.getByRole("img")).toHaveAttribute(
       "src",
-      "//placekitten.com/800/400"
+      getPlaceholderImage("smaller", 0)
     );
   });
 
@@ -140,7 +141,7 @@ describe("Hero", () => {
   it("renders Campaign Hero", () => {
     render(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="campaign"
         heading={<Heading level="h1" id="campaign-hero" text="Hero Campaign" />}
         imageProps={imageProps}
@@ -153,12 +154,12 @@ describe("Hero", () => {
     // @TODO: This test needs to be rewritten to target a pseudo element.
     // expect(screen.getByTestId("hero")).toHaveAttribute(
     //   "style",
-    //   "background-image: url(//placekitten.com/1600/800);"
+    //   `background-image: url(${getPlaceholderImage("smaller", 0)};`
     // );
     expect(screen.getByRole("img")).toBeInTheDocument();
     expect(screen.getByRole("img")).toHaveAttribute(
       "src",
-      "//placekitten.com/800/400"
+      getPlaceholderImage("smaller", 0)
     );
   });
 
@@ -175,7 +176,7 @@ describe("Hero", () => {
     expect(screen.getByRole("img")).toBeInTheDocument();
     expect(screen.getByRole("img")).toHaveAttribute(
       "src",
-      "//placekitten.com/800/400"
+      getPlaceholderImage("smaller", 0)
     );
   });
 
@@ -190,7 +191,7 @@ describe("Hero", () => {
             text="Hero with Custom Colors"
           />
         }
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         foregroundColor="#123456"
         backgroundColor="#654321"
       />
@@ -239,7 +240,7 @@ describe("Hero", () => {
     const warn = jest.spyOn(console, "warn");
     render(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="primary"
         imageProps={{ src: imageProps.src }}
       />
@@ -263,7 +264,7 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="primary"
         heading={heading}
         imageProps={imageProps}
@@ -296,7 +297,7 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="secondary"
         heading={heading}
         imageProps={imageProps}
@@ -358,7 +359,7 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="tertiary"
         heading={heading}
         subHeaderText={otherSubHeaderText}
@@ -377,7 +378,7 @@ describe("Hero", () => {
     );
     const { rerender } = render(
       <Hero
-        backgroundImageSrc="//placekitten.com/g/2400/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="campaign"
         heading={heading}
         imageProps={imageProps}
@@ -410,7 +411,7 @@ describe("Hero", () => {
         heroType="campaign"
         heading={heading}
         subHeaderText={otherSubHeaderText}
-        backgroundImageSrc="//placekitten.com/g/2400/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         locationDetails={<>Some location details.</>}
       />
     );
@@ -497,7 +498,7 @@ describe("Hero", () => {
 
     rerender(
       <Hero
-        backgroundImageSrc="//placekitten.com/g/2400/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="fiftyFifty"
         imageProps={imageProps}
         subHeaderText={otherSubHeaderText}
@@ -516,7 +517,7 @@ describe("Hero", () => {
           heroType="primary"
           heading={<Heading level="h1" id="primary-hero" text="Hero Primary" />}
           subHeaderText="Example Subtitle"
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         />
       )
       .toJSON();
@@ -595,7 +596,7 @@ describe("Hero", () => {
     const campaign = renderer
       .create(
         <Hero
-          backgroundImageSrc="//placekitten.com/g/2400/800"
+          backgroundImageSrc={getPlaceholderImage("smaller", 0)}
           heroType="campaign"
           heading={
             <Heading level="h1" id="campaign-hero" text="Hero Campaign" />
@@ -631,7 +632,7 @@ describe("Hero", () => {
           heroType="primary"
           heading={<Heading level="h1" id="chakra" text="Hero Primary" />}
           subHeaderText="Example Subtitle"
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc={getPlaceholderImage("smaller", 0)}
           p="20px"
           color="ui.error.primary"
         />
@@ -643,7 +644,7 @@ describe("Hero", () => {
           heroType="primary"
           heading={<Heading level="h1" id="props" text="Hero Primary" />}
           subHeaderText="Example Subtitle"
-          backgroundImageSrc="//placekitten.com/1600/800"
+          backgroundImageSrc={getPlaceholderImage("smaller", 0)}
           data-testid="props"
         />
       )
@@ -666,7 +667,7 @@ describe("Hero", () => {
     const ref = React.createRef<HTMLDivElement>();
     const { container } = render(
       <Hero
-        backgroundImageSrc="//placekitten.com/1600/800"
+        backgroundImageSrc={getPlaceholderImage("smaller", 0)}
         heroType="primary"
         heading={<Heading level="h1" id="primary-hero" text="Hero Primary" />}
         ref={ref}