+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `Banner` and `StatusBadge` components to use the `ui.warning.tertiary` color for bertter accessibility in the `"warning"` variant.
 - Updates the styles for the UI colors, text treatment and `Clear all` button in the `TagSet` component.
 - Updates the component category for the `TagSet` component to `"Content Display"`.
+- Updates placeholder images on story pages and tests to use the `getPlaceholderImage` utility function which provides NYPL images.
 
 ## 3.1.0 (April 11, 2024)